+++ src/main/java/kitchenpos/response/TableGroupResponse.java
@@ -1,5 +1,6 @@
 package kitchenpos.response;
 
+import kitchenpos.ordertable.response.OrderTableResponse;
 import kitchenpos.tablegroup.domain.TableGroup;
 
 import java.time.LocalDateTime;
