+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -13,9 +13,9 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.application.TableService;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.request.OrderTableCreateRequest;
-import kitchenpos.request.TableChangeEmptyRequest;
-import kitchenpos.request.TableChangeNumberOfGuestsRequest;
+import kitchenpos.ordertable.request.OrderTableCreateRequest;
+import kitchenpos.ordertable.request.TableChangeEmptyRequest;
+import kitchenpos.ordertable.request.TableChangeNumberOfGuestsRequest;
 import kitchenpos.request.TableGroupCreateRequest;
 import kitchenpos.tablegroup.application.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
